Experimental: Add option to select http4s backend (close #424)
We've recently observed that blaze can have a very different behavioral
semantics than ember. Where the latter might be desirable.
For that reason it is useful to expose a setting allowing us to enable
one or the other.

This change adds `experimental` config section. This section is prone to
backwards-incompatible change or removal.